[NoteTakingApp:1.0.0] Ui Changes

+++ NoteTakingAppCompose/app/src/main/java/com/phoenix/note/screen/Dialog.kt
@@ -1,18 +1,13 @@
 package com.phoenix.note.screen
 
-import androidx.compose.foundation.BorderStroke
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.AlertDialog
 import androidx.compose.material3.Button
-import androidx.compose.material3.ButtonDefaults
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.OutlinedButton
 import androidx.compose.material3.Text
+import androidx.compose.material3.TextButton
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.unit.dp
 import com.phoenix.note.R
 
 /**
@@ -25,36 +20,21 @@ object Dialog {
     fun DeleteDialog(
         onDismiss: () -> Unit,
         onDelete: () -> Unit,
-        title: String,
     ) {
         AlertDialog(
             onDismissRequest = { onDismiss() },
-            title = {
-                Text(text = title)
-            },
+            title = { Text(text = stringResource(R.string.delete_note)) },
             text = { Text(text = stringResource(R.string.delete_message)) },
-            confirmButton = {
-                Button(
-                    onClick = { onDismiss() },
-                    shape = RoundedCornerShape(8.dp),
-                    modifier = Modifier.fillMaxWidth()
-                ) {
+            dismissButton = {
+                TextButton(onClick = { onDismiss() }) {
                     Text(stringResource(R.string.cancel))
                 }
             },
-            dismissButton = {
-                OutlinedButton(
-                    onClick = {
-                        onDelete()
-                        onDismiss()
-                    },
-                    shape = RoundedCornerShape(8.dp),
-                    modifier = Modifier.fillMaxWidth(),
-                    border = BorderStroke(
-                        ButtonDefaults.outlinedButtonBorder.width,
-                        MaterialTheme.colorScheme.error
-                    )
-                ) {
+            confirmButton = {
+                TextButton(onClick = {
+                    onDelete()
+                    onDismiss()
+                }) {
                     Text(
                         stringResource(R.string.delete),
                         color = MaterialTheme.colorScheme.error
@@ -67,33 +47,22 @@ object Dialog {
     @Composable
     fun UnsavedChanges(
         onDismiss: () -> Unit,
-        onConfirm: () -> Unit
+        onConfirm: () -> Unit,
     ) {
         AlertDialog(
             onDismissRequest = { onDismiss() },
+            title = { Text(text = stringResource(R.string.discard_changes))},
             text = { Text(text = stringResource(R.string.exit_message)) },
-            confirmButton = {
-                Button(
-                    onClick = { onDismiss() },
-                    shape = RoundedCornerShape(8.dp),
-                    modifier = Modifier.fillMaxWidth()
-                ) {
+            dismissButton = {
+                TextButton(onClick = { onDismiss() }) {
                     Text(stringResource(R.string.keep))
                 }
             },
-            dismissButton = {
-                OutlinedButton(
-                    onClick = {
-                        onConfirm()
-                        onDismiss()
-                    },
-                    shape = RoundedCornerShape(8.dp),
-                    modifier = Modifier.fillMaxWidth(),
-                    border = BorderStroke(
-                        ButtonDefaults.outlinedButtonBorder.width,
-                        MaterialTheme.colorScheme.error
-                    )
-                ) {
+            confirmButton = {
+                TextButton(onClick = {
+                    onConfirm()
+                    onDismiss()
+                }) {
                     Text(
                         stringResource(R.string.discard),
                         color = MaterialTheme.colorScheme.error
@@ -105,7 +74,7 @@ object Dialog {
 }
 
 sealed interface DialogState {
-    data object None: DialogState
-    data class Delete(val noteId: Int): DialogState
-    data object UnsavedChanges: DialogState
+    data object None : DialogState
+    data class Delete(val noteId: Int) : DialogState
+    data object UnsavedChanges : DialogState
 }
